\name{mwTensor-package}
\alias{mwTensor-package}
\alias{mwTensor}
\docType{package}
\title{
\packageTitle{mwTensor}
}
\description{
\packageDescription{mwTensor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mwTensor}
\packageIndices{mwTensor}
}
\author{
\packageAuthor{mwTensor}

Maintainer: \packageMaintainer{mwTensor}
}
\references{
Andrzej Cichocki et al., (2016). Tensor Networks for Dimensionality Reduction and Large-scale Optimization: Part 1 Low-Rank Tensor Decompositions

Andrzej Cichocki et al., (2015). Tensor Decompositions for Signal Processing Applications, \emph{IEEE SIGNAL PROCESSING MAGAZINE}

Gene H. Golub et al., (2012). Matrix Computation (Johns Hopkins Studies in the Mathematical Sciences), \emph{Johns Hopkins University Press}

Madeleine Udell et al., (2016). Generalized Low Rank Models, \emph{Foundations and Trends in Machine Learning}, 9(1).

Andrzej CICHOCK, et. al., (2009). Nonnegative Matrix and Tensor Factorizations.

A. Hyvarinen. (1999). Fast and Robust Fixed-Point Algorithms for Independent Component Analysis. \emph{IEEE Transactions on Neural Networks}, 10(3), 626-634.

Petros Drineas et al., (2008). Relative-Error CUR Matrix Decompositions, \emph{SIAM Journal on Matrix Analysis and Applications}, 30(2), 844-881.
}

\keyword{ package }
\seealso{\code{\link{mySVD}}, \code{\link{myALS_SVD}}, \code{\link{myNMF}}, \code{\link{myICA}}, \code{\link{myCX}}, \code{\link{MWCA}}, \code{\link{CoupledMWCA}}, \code{\link{checkCoupledMWCA}}, \code{\link{initCoupledMWCA}}, \code{\link{refineFactor}}, \code{\link{MWCAProgram}}, \code{\link{plotTensor3Ds}}}
\examples{
ls("package:mwTensor")
}
